Moved diskutil reporting functions to stat.c
